Update to 151.7909.
Update to commit 1b501fc5acddd423d298d4bbb5e0de16e6a75bf5

+++ chrome/credential_provider/gaiacp/reg_utils.cc
@@ -124,7 +124,7 @@ HRESULT SetMachineRegBinaryInternal(const std::wstring& key_name,
   return S_OK;
 }
 
-std::wstring GetImageRegKeyForSpecificSize(int image_size) {
+std::wstring GetImageRegKeyForSpecificSize(size_t image_size) {
   return kImageRegKey + base::NumberToWString(image_size);
 }
 
@@ -268,7 +268,7 @@ HRESULT GetMachineRegBinaryInternal(const std::wstring& key_name,
 }
 
 HRESULT GetAccountPictureRegString(const std::wstring& user_sid,
-                                   int image_size,
+                                   size_t image_size,
                                    wchar_t* value,
                                    ULONG* length) {
   return GetMachineRegString(GetAccountPictureRegPathForUSer(user_sid),
@@ -278,7 +278,7 @@ HRESULT GetAccountPictureRegString(const std::wstring& user_sid,
 
 // Sets a specific account picture registry key in HKEY_LOCAL_MACHINE
 HRESULT SetAccountPictureRegString(const std::wstring& user_sid,
-                                   int image_size,
+                                   size_t image_size,
                                    const std::wstring& value) {
   return SetMachineRegString(GetAccountPictureRegPathForUSer(user_sid),
                              GetImageRegKeyForSpecificSize(image_size), value);
